[CSL-1800] Move some not csl-specific utility modules into a separate package
