+++ auxx/cardano-sl-auxx.cabal
@@ -44,6 +44,7 @@ library
                      , base58-bytestring
                      , bytestring
                      , canonical-json
+                     , cardano-sl-util
                      , cardano-sl
                      , cardano-sl-core
                      , cardano-sl-db
