+++ auxx/src/Mode.hs
@@ -64,14 +64,13 @@ import           Pos.Ssc.Types (HasSscContext (..), SscBlock)
 import           Pos.Txp (MempoolExt, MonadTxpLocal (..), txNormalize, txProcessTransaction,
                           txProcessTransactionNoLock)
 import           Pos.Txp.DB.Utxo (getFilteredUtxo)
-import           Pos.Util (Some (..))
+import           Pos.Util (HasLens (..), Some (..), postfixLFields)
 import           Pos.Util.CompileInfo (HasCompileInfo, withCompileInfo)
 import           Pos.Util.JsonLog (HasJsonLogConfig (..))
 import           Pos.Util.LoggerName (HasLoggerName' (..))
 import qualified Pos.Util.OutboundQueue as OQ.Reader
 import           Pos.Util.TimeWarp (CanJsonLog (..))
 import           Pos.Util.UserSecret (HasUserSecret (..))
-import           Pos.Util.Util (HasLens (..), postfixLFields)
 import           Pos.WorkMode (EmptyMempoolExt, RealMode, RealModeContext (..))
 
 -- | Command execution context.
