+++ core/Pos/Util/Util.hs
@@ -1,21 +1,13 @@
 {-# LANGUAGE CPP             #-}
 {-# LANGUAGE ConstraintKinds #-}
-{-# LANGUAGE GADTs           #-}
 {-# LANGUAGE PolyKinds       #-}
 {-# LANGUAGE RankNTypes      #-}
 {-# LANGUAGE TypeFamilies    #-}
 
 module Pos.Util.Util
        (
-       -- * Existential types
-         Some(..)
-       , Some1(..)
-       , applySome
-       , liftLensSome
-       , liftGetterSome
-
        -- * Something
-       , Sign (..)
+         Sign (..)
        , maybeThrow
        , eitherToFail
        , eitherToThrow
@@ -27,13 +19,6 @@ module Pos.Util.Util
        , median
        , (<//>)
        , divRoundUp
-
-       -- * Lenses
-       , _neHead
-       , _neTail
-       , _neLast
-       , postfixLFields
-       , postfixLFields2
 
        -- * Ether
        , ether
@@ -94,11 +79,9 @@ module Pos.Util.Util
        ) where
 
 import           Universum
-import           Unsafe (unsafeInit, unsafeLast)
 
 import           Control.Concurrent (myThreadId, threadDelay)
-import           Control.Lens (ALens', Getter, Getting, Iso', LensRules, cloneLens, coerced,
-                               foldMapOf, lensField, lensRules, mappingNamer, to, ( # ))
+import           Control.Lens (Getting, Iso', coerced, foldMapOf, ( # ))
 import           Control.Monad.Base (MonadBase)
 import qualified Control.Monad.Catch as MC
 import           Control.Monad.Morph (MFunctor (..))
@@ -144,44 +127,6 @@ import           System.IO (hClose, openTempFile)
 import           System.Wlog (CanLog, HasLoggerName (..), LoggerNameBox (..))
 import qualified Test.QuickCheck as QC
 import           Test.QuickCheck.Monadic (PropertyM (..))
-
-----------------------------------------------------------------------------
--- Some
-----------------------------------------------------------------------------
-
--- | Turn any constraint into an existential type! Example:
---
--- @
--- foo :: Some Show -> String
--- foo (Some s) = show s
--- @
-data Some c where
-    Some :: c a => a -> Some c
-
-instance Show (Some Show) where
-    show (Some s) = show s
-
--- | Like 'Some', but for @* -> *@ types – for instance, @Some1 Functor ()@
-data Some1 c a where
-    Some1 :: c f => f a -> Some1 c a
-
-instance Functor (Some1 Functor) where
-    fmap f (Some1 x) = Some1 (fmap f x)
-
--- | Apply a function requiring the constraint
-applySome :: (forall a. c a => a -> r) -> (Some c -> r)
-applySome f (Some x) = f x
-
--- | Turn a lens into something operating on 'Some'. Useful for many types
--- like 'HasDifficulty', 'IsHeader', etc.
-liftLensSome :: (forall a. c a => ALens' a b)
-             -> Lens' (Some c) b
-liftLensSome l =
-    \f (Some a) -> Some <$> cloneLens l f a
-
--- | Like 'liftLensSome', but for getters.
-liftGetterSome :: (forall a. c a => Getting b a b) -> Getter (Some c) b
-liftGetterSome l = \f (Some a) -> Some <$> to (view l) f a
 
 ----------------------------------------------------------------------------
 -- Instances
@@ -445,27 +390,6 @@ inAssertMode _ = pure ()
 divRoundUp :: Integral a => a -> a -> a
 divRoundUp a b = (a + b - 1) `div` b
 
-----------------------------------------------------------------------------
--- Lenses
-----------------------------------------------------------------------------
-
--- | Lens for the head of 'NonEmpty'.
---
--- We can't use '_head' because it doesn't work for 'NonEmpty':
--- <https://github.com/ekmett/lens/issues/636#issuecomment-213981096>.
--- Even if we could though, it wouldn't be a lens, only a traversal.
-_neHead :: Lens' (NonEmpty a) a
-_neHead f (x :| xs) = (:| xs) <$> f x
-
--- | Lens for the tail of 'NonEmpty'.
-_neTail :: Lens' (NonEmpty a) [a]
-_neTail f (x :| xs) = (x :|) <$> f xs
-
--- | Lens for the last element of 'NonEmpty'.
-_neLast :: Lens' (NonEmpty a) a
-_neLast f (x :| []) = (:| []) <$> f x
-_neLast f (x :| xs) = (\y -> x :| unsafeInit xs ++ [y]) <$> f (unsafeLast xs)
-
 -- | Print splices generated by a TH splice (the printing will happen during
 -- compilation, as a GHC warning). Useful for debugging.
 --
@@ -480,12 +404,6 @@ dumpSplices x = do
     let code = Prelude.lines (TH.pprint ds)
     TH.reportWarning ("\n" ++ Prelude.unlines (map ("    " ++) code))
     return ds
-
-postfixLFields :: LensRules
-postfixLFields = lensRules & lensField .~ mappingNamer (\s -> [s++"_L"])
-
-postfixLFields2 :: LensRules
-postfixLFields2 = lensRules & lensField .~ mappingNamer (\s -> [s++"L"])
 
 -- | Count elements in a list.
 histogram :: forall a. Ord a => [a] -> Map a Int
