+++ core/cardano-sl-core.cabal
@@ -128,19 +128,10 @@ library
 
                        Pos.Util.Arbitrary
                        Pos.Util.CompileInfo
-                       Pos.Util.Concurrent
-                       Pos.Util.Concurrent.LockedTVar
-                       Pos.Util.Concurrent.PriorityLock
-                       Pos.Util.Concurrent.RWLock
-                       Pos.Util.Concurrent.RWVar
                        Pos.Util.Chrono
-                       Pos.Util.Justify
                        Pos.Util.Limits
-                       Pos.Util.LRU
                        Pos.Util.Mnemonics
                        Pos.Util.Modifier
-                       Pos.Util.Future
-                       Pos.Util.Queue
                        Pos.Util.Util
 
   other-modules:
@@ -157,6 +148,7 @@ library
                      , bytestring
                      , canonical-json
                      , cardano-crypto
+                     , cardano-sl-util
                      , cborg
                      , cereal
                      , concurrent-extra
