+++ infra/cardano-sl-infra.cabal
@@ -120,6 +120,7 @@ library
                       , base64-bytestring
                       , binary
                       , bytestring
+                      , cardano-sl-util
                       , cardano-sl-core
                       , cardano-sl-db
                       , cardano-report-server >= 0.3.0
