+++ lib/cardano-sl.cabal
@@ -250,6 +250,7 @@ library
                       , canonical-json
                       , cardano-crypto
                       , cardano-report-server >= 0.2.1
+                      , cardano-sl-util
                       , cardano-sl-core
                       , cardano-sl-db
                       , cardano-sl-infra
