+++ lib/src/Pos/Launcher/Mode.hs
@@ -50,7 +50,7 @@ import           Pos.Slotting.Impl.Sum (SlottingContextSum, currentTimeSlottingS
 import           Pos.Slotting.MemState (MonadSlotsData)
 import           Pos.Ssc.Types (SscBlock)
 import           Pos.Util (Some (..))
-import           Pos.Util.Util (postfixLFields)
+import           Pos.Util.Lens (postfixLFields)
 
 -- The fields are lazy on purpose: this allows using them with
 -- futures.
@@ -119,4 +119,3 @@ instance (HasConfiguration, HasInfraConfiguration, MonadSlotsData ctx InitMode) 
     getCurrentSlotBlocking   = getCurrentSlotBlockingSum
     getCurrentSlotInaccurate = getCurrentSlotInaccurateSum
     currentTimeSlotting      = currentTimeSlottingSum
-
