+++ lib/src/Pos/Ssc/Util.hs
@@ -11,7 +11,7 @@ import           Pos.Core (HasConfiguration, IsGenesisHeader, IsMainHeader, gbBo
 import           Pos.Core.Block (Block, GenesisBlock, MainBlock, mbSscPayload)
 import           Pos.Core.Ssc (SscPayload)
 import           Pos.Ssc.Types (SscBlock)
-import           Pos.Util (Some (..))
+import           Pos.Util.Some (Some (..))
 
 -- [CSL-1156] Totally need something more elegant
 toSscBlock
