+++ ssc/cardano-sl-ssc.cabal
@@ -71,6 +71,7 @@ library
                      , array
                      , base
                      , bytestring
+                     , cardano-sl-util
                      , cardano-sl-core
                      , cardano-sl-db
                      , cardano-sl-infra
