+++ update/cardano-sl-update.cabal
@@ -86,6 +86,7 @@ library
                      , base
                      , binary
                      , bytestring
+                     , cardano-sl-util
                      , cardano-sl-core
                      , cardano-sl-db
                      , cardano-sl-infra
