+++ util/cardano-sl-util.cabal
@@ -1,7 +1,9 @@
-name:                cardano-sl-db
+name:                cardano-sl-util
 version:             1.0.3
-synopsis:            Cardano SL - basic DB interfaces
-description:         Cardano SL - basic DB interfaces
+synopsis:            Cardano SL - general utilities
+description:         This packages contains utility functions not specific
+                     to Cardano SL which extend 3rd party libraries or implement
+                     something from scratch.
 license:             MIT
 license-file:        LICENSE
 author:              Serokell
@@ -11,47 +13,56 @@ category:            Currency
 build-type:          Simple
 cabal-version:       >=1.10
 
-library
-  exposed-modules:     Pos.DB
-                       Pos.DB.BatchOp
-                       Pos.DB.Class
-                       Pos.DB.Error
-                       Pos.DB.Functions
-                       Pos.DB.GState.Common
-                       Pos.DB.GState.Stakes
-                       Pos.DB.Misc.Common
-                       Pos.DB.Pure
-                       Pos.DB.Sum
-                       Pos.DB.Rocks
-                       Pos.DB.Rocks.Functions
-                       Pos.DB.Rocks.Types
+flag asserts
+  default:     True
+  manual:      True
 
-  build-depends:       base
+library
+  exposed-modules:
+                       Pos.Util.Concurrent
+                       Pos.Util.Concurrent.LockedTVar
+                       Pos.Util.Concurrent.PriorityLock
+                       Pos.Util.Concurrent.RWLock
+                       Pos.Util.Concurrent.RWVar
+
+                       Pos.Util.Future
+                       Pos.Util.Justify
+                       Pos.Util.Lens
+                       Pos.Util.LRU
+                       Pos.Util.Queue
+                       Pos.Util.Some
+
+  other-modules:
+
+  build-depends:       QuickCheck
+                     , autoexporter
+                     , base
                      , bytestring
-                     , cardano-sl-util
-                     , cardano-sl-core
-                     , concurrent-extra
-                     , conduit
                      , containers
+                     , concurrent-extra
                      , data-default
-                     , directory
-                     , ether >= 0.5
-                     , filepath
+                     , deepseq
+                     , exceptions
                      , formatting
+                     , hashable
                      , lens
-                     , log-warper
-                     , monad-control
+                     , log-warper >= 1.0.3
+                     , lrucache
                      , mmorph
                      , mtl
-                     , node-sketch
-                     , resourcet
-                     , rocksdb-haskell >= 1.0.0
+                     , random
+                     , semigroups
                      , serokell-util
+                     , stm
+                     , template-haskell
+                     , text
                      , text-format
                      , transformers
                      , transformers-base
                      , transformers-lift
                      , universum
+                     , unordered-containers
+                     , vector
 
   default-language:    Haskell2010
 
@@ -84,3 +95,5 @@ library
 
   build-tools: cpphs >= 1.19
   ghc-options: -pgmP cpphs -optP --cpp
+  if flag(asserts)
+    cpp-options: -DASSERTS_ON
