+++ wallet/test/Test/Pos/Wallet/Web/Util.hs
@@ -45,11 +45,11 @@ import           Pos.Generator.Block (genBlocks)
 import           Pos.Launcher (HasConfigurations)
 import           Pos.StateLock (Priority (..), modifyStateLock)
 import           Pos.Txp.Toil (Utxo)
+import           Pos.Util (HasLens (..), _neLast)
 import           Pos.Util.Chrono (OldestFirst (..))
 import           Pos.Util.CompileInfo (HasCompileInfo)
 import           Pos.Util.Servant (encodeCType)
 import           Pos.Util.UserSecret (mkGenesisWalletUserSecret)
-import           Pos.Util.Util (HasLens (..), _neLast)
 import           Pos.Wallet.Web.ClientTypes (Addr, CId, Wal, encToCId)
 import           Pos.Wallet.Web.Methods.Restore (importWalletDo)
